Moved the dummy DataGrid data to a separate file/folder
